[UnitTests] Try and pacify gcc-5
This looks like a defect in gcc-5 where it chooses a constexpr
constructor from the initializer-list that it considers to be explicit.

I've tried to reproduce but I can't install anything prior to gcc-6 easily
on my system, and that doesn't have the error. So this is speculative
pacification.

Reported by Steven Wan.

git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@373820 91177308-0d34-0410-b5e6-96231b3b80d8